Split files into list/ and listItem/
