+++ src/app/listItem/create.v1.controller.ts
@@ -3,7 +3,7 @@ import { JsonController, Post, Body, Res } from 'routing-controllers'
 import { IsNotEmpty, IsString, IsArray, MaxLength } from 'class-validator'
 import { Response } from 'koa'
 import HttpStatus from 'http-status-codes'
-import ListItemService from '../domain/listItem.service'
+import ListItemService from '../../domain/listItem/listItem.service'
 
 class CreateRequestParamsDTO {
   @IsNotEmpty()
