+++ src/app/listItem/finish.v1.controller.ts
@@ -1,3 +1,4 @@
+import { AssertionError } from 'assert'
 import { singleton } from 'tsyringe'
 import {
   JsonController,
@@ -8,9 +9,8 @@ import {
 } from 'routing-controllers'
 import { IsUUID } from 'class-validator'
 import HttpStatus from 'http-status-codes'
-import { AssertionError } from 'assert'
-import ListItemService from '../domain/listItem.service'
-import ListItemCreated from '../domain/listItemCreated.event'
+import ListItemService from '../../domain/listItem/listItem.service'
+import ListItemCreated from '../../domain/listItem/listItemCreated.event'
 
 class FinishRequestParamsDTO {
   @IsUUID()
