+++ src/app/listItem/listLabels.v1.controller.ts
@@ -1,6 +1,6 @@
 import { singleton } from 'tsyringe'
 import { JsonController, Get } from 'routing-controllers'
-import ListItemService from '../domain/listItem.service'
+import ListItemService from '../../domain/listItem/listItem.service'
 
 class ListLabelsResponseDTO {
   data: string[]
@@ -8,7 +8,7 @@ class ListLabelsResponseDTO {
 
 @singleton()
 @JsonController('/v1/lists/list-items-labels')
-export default class ListListItemLabelsV1Controller {
+export default class ListItemLabelsV1Controller {
   constructor(private service: ListItemService) {}
 
   @Get('/')
