+++ src/app/listItem/listUnfinished.v1.controller.spec.ts
@@ -4,13 +4,13 @@ import { expect } from 'chai'
 import request from 'supertest'
 import HttpStatus from 'http-status-codes'
 import { uuid } from 'uuidv4'
-import ListItemFinished from '../domain/listItemFinished.event'
-import { createApp, cleanUpApp } from '../app'
-import ConfigService from '../domain/config.service'
-import ListItemStore from '../infra/listItem.store'
-import ListItemCreated from '../domain/listItemCreated.event'
-import EventMockStore from '../infra/event.store.mock'
-import EventStore from '../infra/event.store'
+import { createApp, cleanUpApp } from '../../app'
+import ConfigService from '../../domain/config.service'
+import ListItemFinished from '../../domain/listItem/listItemFinished.event'
+import ListItemCreated from '../../domain/listItem/listItemCreated.event'
+import EventMockStore from '../../infra/event.store.mock'
+import EventStore from '../../infra/event.store'
+import ListItemStore from '../../infra/listItem/listItem.store'
 
 describe('ListUnfinishedListItemV1Controller', () => {
   let server: Server
