+++ src/infra/list/list.store.ts
@@ -1,8 +1,8 @@
 import { singleton } from 'tsyringe'
 import { EventEmitter } from 'typed-ts-events'
-import BaseStore from './base.store'
-import AutoLoadableStore from './autoLoadableStore.interface'
-import EventStore, { Event } from './event.store'
+import BaseStore from '../base.store'
+import AutoLoadableStore from '../autoLoadableStore.interface'
+import EventStore, { Event } from '../event.store'
 
 @singleton()
 export default class ListStore extends BaseStore implements AutoLoadableStore {
